ARK-004 Repository 구조 변경
